Make it work with single files

+++ test.py
@@ -3,7 +3,8 @@
 
 TESTS_DIR = 'tests_data'
 EXAMPLES = {'ejemplo': ['--no-split'],
-            'ejemplo_2': []}
+            'ejemplo_2': [],
+            'ejemplo_single': []}
 
 import glob
 import shutil
@@ -20,7 +21,9 @@ def run_and_assert(cmd):
 
 def clean():
     for ex in EXAMPLES.keys():
-        shutil.rmtree(os.path.join(TESTS_DIR, ex))
+        dir = os.path.join(TESTS_DIR, ex)
+        if os.path.isdir(dir):
+            shutil.rmtree(dir)
 
 
 def run_tests():
@@ -28,7 +31,8 @@ def run_tests():
         new_dir = os.path.join(TESTS_DIR, ex)
         os.mkdir(new_dir)
         for f in glob.glob(os.path.join(ex, '*.txt')) + [os.path.join(ex, 'config.py')]:
-            shutil.copy(f, new_dir)
+            if os.path.isfile(f):
+                shutil.copy(f, new_dir)
         cmd = ['python', 'automagica.py', '--only-tex'] + EXAMPLES[ex] + [os.path.join(TESTS_DIR, ex)]
         run_and_assert(cmd)
 
@@ -36,11 +40,14 @@ def run_tests():
     run_and_assert(cmd)
     cmd = ['diff', 'tests_data/ejemplo_2/sueltos.tex', 'tests_data/sueltos.tex']
     run_and_assert(cmd)
+    cmd = ['diff', 'tests_data/ejemplo_single/default.tex', 'tests_data/default.tex']
+    run_and_assert(cmd)
     clean()
 
 
 if __name__ == '__main__':
     try:
         run_tests()
-    except:
+    except Exception as e:
+        print e
         clean()
